BUG: Corrigindo Anotações do service

+++ src/main/java/br/com/alura/comex/model/DadosNovoProduto.java
@@ -3,9 +3,10 @@ package br.com.alura.comex.model;
 import jakarta.validation.constraints.Min;
 import org.hibernate.validator.constraints.Length;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
-public record DadosNovoProduto(@Length(min = 5) String nomeProduto, double preco, @Length(min = 1) int quantidade,
+public record DadosNovoProduto(@Length(min = 5) String nomeProduto, BigDecimal preco, @Length(min = 1) int quantidade,
                                Long idCategoria) {
 
     public Produto toEntity(Categoria categoria) {
